requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
entrypoints==0.4
fastapi==0.88.0
ffmpy==0.3.0
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gradio==3.15.0
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
huggingface-hub==0.11.1
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.3
importlib-metadata==5.2.0
importlib-resources==5.10.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.3
ninja==1.11.1
numpy==1.24.0
oauthlib==3.2.2
opencv-python==4.6.0.66
orjson==3.8.3
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
pydantic==1.10.2
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyspng==0.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toolz==0.12.0
torch==1.7.1+cu110
tqdm==4.64.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
